ci(config,readme): add nvmrc, npmrc & engine locking config and updated ReadMe of client app
Added the configurations to prevent user for utilising npm and node version other than 18. This ensures that all developers working on this projects uses same tools to develop the project.

Updated the README file of the client application to inform user/developer about the project and give them basic instructions on how to run it